+++ android/src/main/kotlin/com/fluttercandies/photo_manager/constant/Methods.kt
@@ -2,14 +2,61 @@ package com.fluttercandies.photo_manager.constant
 
 class Methods {
     companion object {
+        // Not need permission methods
         const val log = "log"
         const val openSetting = "openSetting"
         const val forceOldAPI = "forceOldApi"
         const val systemVersion = "systemVersion"
         const val clearFileCache = "clearFileCache"
         const val releaseMemoryCache = "releaseMemoryCache"
+        const val ignorePermissionCheck = "ignorePermissionCheck"
 
+        fun isNotNeedPermissionMethod(method: String): Boolean {
+            return method in arrayOf(
+                log,
+                openSetting,
+                forceOldAPI,
+                systemVersion,
+                clearFileCache,
+                releaseMemoryCache,
+                ignorePermissionCheck,
+            )
+        }
+        // Not need permission methods end
+
+        // About permission start
         const val requestPermissionExtend = "requestPermissionExtend"
+        const val presentLimited = "presentLimited"
+
+        fun isPermissionMethod(method: String): Boolean {
+            return method in arrayOf(
+                requestPermissionExtend,
+                presentLimited,
+            )
+        }
+        // About permission end
+
+        /// Have [requestType] start
+        const val fetchPathProperties = "fetchPathProperties"
+        const val getAssetPathList = "getAssetPathList"
+        const val getAssetListPaged = "getAssetListPaged"
+        const val getAssetListRange = "getAssetListRange"
+        const val getAssetCount = "getAssetCount"
+        const val getAssetsByRange = "getAssetsByRange"
+
+        val haveRequestTypeMethods = arrayOf(
+            fetchPathProperties,
+            getAssetPathList,
+            getAssetListPaged,
+            getAssetListRange,
+            getAssetCount,
+            getAssetsByRange,
+        )
+
+        fun isHaveRequestTypeMethod(method: String): Boolean {
+            return method in haveRequestTypeMethods
+        }
+        /// Have [requestType] end
 
         const val getThumbnail = "getThumb"
         const val requestCacheAssetsThumbnail = "requestCacheAssetsThumb"
@@ -31,20 +78,22 @@ class Methods {
         const val removeNoExistsAssets = "removeNoExistsAssets"
         const val getColumnNames = "getColumnNames"
 
-        const val getAssetCount = "getAssetCount"
-        const val getAssetsByRange = "getAssetsByRange"
-
-        /// Below methods have [RequestType] params, thus permissions are required for Android 13.
-        const val fetchPathProperties = "fetchPathProperties"
-        const val getAssetPathList = "getAssetPathList"
-        const val getAssetListPaged = "getAssetListPaged"
-        const val getAssetListRange = "getAssetListRange"
-
-        val android13PermissionMethods = arrayOf(
-            fetchPathProperties,
-            getAssetPathList,
-            getAssetListPaged,
-            getAssetListRange,
+        val needMediaLocationMethods = arrayOf(
+            getLatLng,
+            getFullFile,
+            getOriginBytes,
         )
+
+        fun isNeedMediaLocationMethod(method: String): Boolean {
+            return method in needMediaLocationMethods
+        }
+
+        fun otherMethods(method: String): Boolean {
+            return (isNotNeedPermissionMethod(method) ||
+                    isPermissionMethod(method) ||
+                    isHaveRequestTypeMethod(method) ||
+                    isNeedMediaLocationMethod(method))
+                .not()
+        }
     }
 }