+++ android/src/main/kotlin/com/fluttercandies/photo_manager/core/utils/AndroidQDBUtils.kt
@@ -15,8 +15,8 @@ import androidx.exifinterface.media.ExifInterface
 import com.fluttercandies.photo_manager.core.PhotoManager
 import com.fluttercandies.photo_manager.core.cache.ScopedCache
 import com.fluttercandies.photo_manager.core.entity.AssetEntity
-import com.fluttercandies.photo_manager.core.entity.filter.FilterOption
 import com.fluttercandies.photo_manager.core.entity.AssetPathEntity
+import com.fluttercandies.photo_manager.core.entity.filter.FilterOption
 import com.fluttercandies.photo_manager.util.LogUtils
 import java.io.ByteArrayOutputStream
 import java.io.File
@@ -44,7 +44,7 @@ object AndroidQDBUtils : IDBUtils {
         val selections =
             "$BUCKET_ID IS NOT NULL $where"
 
-        val cursor = context.contentResolver.query(
+        val cursor = context.contentResolver.logQuery(
             allUri,
             IDBUtils.storeBucketKeys,
             selections,
@@ -90,7 +90,7 @@ object AndroidQDBUtils : IDBUtils {
         val selections =
             "$BUCKET_ID IS NOT NULL $where"
 
-        val cursor = context.contentResolver.query(
+        val cursor = context.contentResolver.logQuery(
             allUri,
             IDBUtils.storeBucketKeys,
             selections,
@@ -154,12 +154,12 @@ object AndroidQDBUtils : IDBUtils {
             "$BUCKET_ID = ? $where"
         }
         val sortOrder = getSortOrder(page * size, size, option)
-        val cursor = context.contentResolver.query(
-                allUri,
-                keys(),
-                selection,
-                args.toTypedArray(),
-                sortOrder
+        val cursor = context.contentResolver.logQuery(
+            allUri,
+            keys(),
+            selection,
+            args.toTypedArray(),
+            sortOrder
         ) ?: return list
         cursor.use {
             cursorWithRange(it, page * size, size) { cursor ->
@@ -194,7 +194,7 @@ object AndroidQDBUtils : IDBUtils {
         }
         val pageSize = end - start
         val sortOrder = getSortOrder(start, pageSize, option)
-        val cursor = context.contentResolver.query(
+        val cursor = context.contentResolver.logQuery(
             allUri,
             keys(),
             selection,
@@ -215,23 +215,23 @@ object AndroidQDBUtils : IDBUtils {
 
     override fun keys(): Array<String> {
         return (IDBUtils.storeImageKeys + IDBUtils.storeVideoKeys + IDBUtils.typeKeys + arrayOf(
-                RELATIVE_PATH
+            RELATIVE_PATH
         )).distinct().toTypedArray()
     }
 
     override fun getAssetEntity(
-            context: Context,
-            id: String,
-            checkIfExists: Boolean
+        context: Context,
+        id: String,
+        checkIfExists: Boolean
     ): AssetEntity? {
         val selection = "$_ID = ?"
         val args = arrayOf(id)
-        val cursor = context.contentResolver.query(
-                allUri,
-                keys(),
-                selection,
-                args,
-                null
+        val cursor = context.contentResolver.logQuery(
+            allUri,
+            keys(),
+            selection,
+            args,
+            null
         ) ?: return null
         cursor.use {
             return if (it.moveToNext()) it.toAssetEntity(context, checkIfExists)
@@ -260,7 +260,7 @@ object AndroidQDBUtils : IDBUtils {
 
         val selection =
             "$BUCKET_ID IS NOT NULL $where $idSelection"
-        val cursor = context.contentResolver.query(
+        val cursor = context.contentResolver.logQuery(
             allUri,
             IDBUtils.storeBucketKeys,
             selection, args.toTypedArray(),
@@ -351,7 +351,7 @@ object AndroidQDBUtils : IDBUtils {
         }
 
         val cr = context.contentResolver
-        val cursor = cr.query(
+        val cursor = cr.logQuery(
             allUri,
             copyKeys.toTypedArray() + arrayOf(RELATIVE_PATH),
             idSelection,
@@ -421,7 +421,7 @@ object AndroidQDBUtils : IDBUtils {
             Log.i(TAG, "The removeAllExistsAssets is starting.")
             val removedList = ArrayList<String>()
             val cr = context.contentResolver
-            val cursor = cr.query(
+            val cursor = cr.logQuery(
                 allUri,
                 arrayOf(
                     BaseColumns._ID,
@@ -479,7 +479,7 @@ object AndroidQDBUtils : IDBUtils {
 
     private fun getRelativePath(context: Context, galleryId: String): String? {
         val cr = context.contentResolver
-        val cursor = cr.query(
+        val cursor = cr.logQuery(
             allUri,
             arrayOf(BUCKET_ID, RELATIVE_PATH),
             "$BUCKET_ID = ?",
@@ -496,7 +496,7 @@ object AndroidQDBUtils : IDBUtils {
 
     override fun getSomeInfo(context: Context, assetId: String): Pair<String, String?>? {
         val cr = context.contentResolver
-        val cursor = cr.query(
+        val cursor = cr.logQuery(
             allUri,
             arrayOf(BUCKET_ID, RELATIVE_PATH),
             "$_ID = ?",