+++ android/src/main/kotlin/com/fluttercandies/photo_manager/permission/PermissionsListener.kt
@@ -1,6 +1,12 @@
 package com.fluttercandies.photo_manager.permission
 
 interface PermissionsListener {
-    fun onGranted()
-    fun onDenied(deniedPermissions: MutableList<String>, grantedPermissions: MutableList<String>)
+
+    fun onGranted(needPermissions: MutableList<String>)
+
+    fun onDenied(
+        deniedPermissions: MutableList<String>,
+        grantedPermissions: MutableList<String>,
+        needPermissions: MutableList<String>
+    )
 }