+++ lib/src/internal/image_provider.dart
@@ -2,6 +2,7 @@
 // Use of this source code is governed by an Apache license that can be found
 // in the LICENSE file.
 
+import 'dart:async';
 import 'dart:io';
 import 'dart:typed_data' as typed_data;
 import 'dart:ui' as ui;
@@ -13,6 +14,8 @@ import '../types/entity.dart';
 import '../types/thumbnail.dart';
 import 'constants.dart';
 import 'enums.dart';
+
+final _providerLocks = <AssetEntityImageProvider, Completer<ui.Codec>>{};
 
 /// The [ImageProvider] that handles [AssetEntity].
 ///
@@ -91,59 +94,73 @@ class AssetEntityImageProvider extends ImageProvider<AssetEntityImageProvider> {
   Future<ui.Codec> _loadAsync(
     AssetEntityImageProvider key,
     DecoderCallback decode, // ignore: deprecated_member_use
-  ) async {
-    try {
-      assert(key == this);
-      if (key.entity.type == AssetType.audio ||
-          key.entity.type == AssetType.other) {
-        throw UnsupportedError(
-          'Image data for the ${key.entity.type} is not supported.',
-        );
-      }
-
-      // Define the image type.
-      final ImageFileType type;
-      if (key.imageFileType == ImageFileType.other) {
-        // Assume the title is invalid here, try again with the async getter.
-        type = _getType(await key.entity.titleAsync);
-      } else {
-        type = key.imageFileType;
-      }
-
-      typed_data.Uint8List? data;
-      if (isOriginal) {
-        if (key.entity.type == AssetType.video) {
-          data = await key.entity.thumbnailData;
-        } else if (type == ImageFileType.heic) {
-          data = await (await key.entity.file)?.readAsBytes();
-        } else {
-          data = await key.entity.originBytes;
-        }
-      } else {
-        data = await key.entity.thumbnailDataWithOption(
-          _thumbOption(thumbnailSize!),
-        );
-      }
-      if (data == null) {
-        throw StateError('The data of the entity is null: $entity');
-      }
-      return decode(data);
-    } catch (e, s) {
-      if (kDebugMode) {
-        FlutterError.presentError(
-          FlutterErrorDetails(
-            exception: e,
-            stack: s,
-            library: PMConstants.libraryName,
-          ),
-        );
-      }
-      // Depending on where the exception was thrown, the image cache may not
-      // have had a chance to track the key in the cache at all.
-      // Schedule a microtask to give the cache a chance to add the key.
-      Future<void>.microtask(() => _evictCache(key));
-      rethrow;
+  ) {
+    if (_providerLocks.containsKey(key)) {
+      return _providerLocks[key]!.future;
     }
+    final lock = Completer<ui.Codec>();
+    _providerLocks[key] = lock;
+    Future(() async {
+      try {
+        assert(key == this);
+        if (key.entity.type == AssetType.audio ||
+            key.entity.type == AssetType.other) {
+          throw UnsupportedError(
+            'Image data for the ${key.entity.type} is not supported.',
+          );
+        }
+
+        // Define the image type.
+        final ImageFileType type;
+        if (key.imageFileType == ImageFileType.other) {
+          // Assume the title is invalid here, try again with the async getter.
+          type = _getType(await key.entity.titleAsync);
+        } else {
+          type = key.imageFileType;
+        }
+
+        typed_data.Uint8List? data;
+        if (isOriginal) {
+          if (key.entity.type == AssetType.video) {
+            data = await key.entity.thumbnailData;
+          } else if (type == ImageFileType.heic) {
+            data = await (await key.entity.file)?.readAsBytes();
+          } else {
+            data = await key.entity.originBytes;
+          }
+        } else {
+          data = await key.entity.thumbnailDataWithOption(
+            _thumbOption(thumbnailSize!),
+          );
+        }
+        if (data == null) {
+          throw StateError('The data of the entity is null: $entity');
+        }
+        return decode(data);
+      } catch (e, s) {
+        if (kDebugMode) {
+          FlutterError.presentError(
+            FlutterErrorDetails(
+              exception: e,
+              stack: s,
+              library: PMConstants.libraryName,
+            ),
+          );
+        }
+        // Depending on where the exception was thrown, the image cache may not
+        // have had a chance to track the key in the cache at all.
+        // Schedule a microtask to give the cache a chance to add the key.
+        Future<void>.microtask(() => _evictCache(key));
+        rethrow;
+      }
+    }).then((codec) {
+      lock.complete(codec);
+    }).catchError((e, s) {
+      lock.completeError(e, s);
+    }).whenComplete(() {
+      _providerLocks.remove(key);
+    });
+    return lock.future;
   }
 
   ThumbnailOption _thumbOption(ThumbnailSize size) {